backendProject/requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-cors-headers==2.5.2
django-rest-auth==0.9.3
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
httpie==1.0.2
idna==2.8
Markdown==3.0.1
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pygments==2.3.1
PyJWT==1.7.1
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.1
